requirements.txt
openai==1.64.0
langchain==0.1.11
langchain-openai==0.0.8
python-dotenv==1.0.1